+++ lib/src/command_runner.dart
@@ -2,10 +2,10 @@ import 'package:args/args.dart' show ArgResults;
 import 'package:args/command_runner.dart' show CommandRunner, UsageException;
 import 'package:logger/logger.dart' show Logger, ProductionFilter;
 
+import '../version.dart';
 import 'commands/command_status.dart';
 import 'commands/create.dart';
 import 'util/printer.dart';
-import '../version.dart';
 
 class AtCommandRunner extends CommandRunner<CommandStatus> {
   final Logger _logger;
