--- lib/src/commands/constants/dart_keywords.dart
@@ -1,73 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// dart keywords for package name validation
-const Set<String> dartKeywords = <String>{
-  'abstract',
-  'as',
-  'assert',
-  'async',
-  'await',
-  'break',
-  'case',
-  'catch',
-  'class',
-  'const',
-  'continue',
-  'covariant',
-  'default',
-  'deferred',
-  'do',
-  'dynamic',
-  'else',
-  'enum',
-  'export',
-  'extends',
-  'extension',
-  'external',
-  'factory',
-  'false',
-  'final',
-  'finally',
-  'for',
-  'function',
-  'get',
-  'hide',
-  'if',
-  'implements',
-  'import',
-  'in',
-  'inout',
-  'interface',
-  'is',
-  'late',
-  'library',
-  'mixin',
-  'native',
-  'new',
-  'null',
-  'of',
-  'on',
-  'operator',
-  'out',
-  'part',
-  'patch',
-  'required',
-  'rethrow',
-  'return',
-  'set',
-  'show',
-  'source',
-  'static',
-  'super',
-  'switch',
-  'sync',
-  'this',
-  'throw',
-  'true',
-  'try',
-  'typedef',
-  'var',
-  'void',
-  'while',
-  'with',
-  'yield',
-};
