+++ lib/src/commands/create.dart
@@ -1,17 +1,17 @@
-import 'package:at_app/src/util/namespace.dart';
+import '../util/namespace.dart';
 import 'package:logger/logger.dart' show Logger, ProductionFilter;
 import 'package:path/path.dart' show join, relative;
 
 import '../util/cache_package.dart';
-import '../util/cli/flutter.dart';
-import '../util/exceptions/android_build_exception.dart';
-import '../util/exceptions/env_exception.dart';
-import '../util/exceptions/package_exception.dart';
-import '../util/exceptions/template_exception.dart';
+import '../cli/flutter_cli.dart';
+import '../models/exceptions/android_build_exception.dart';
+import '../models/exceptions/env_exception.dart';
+import '../models/exceptions/package_exception.dart';
+import '../models/exceptions/template_exception.dart';
 import '../util/printer.dart';
 import '../util/template_manager.dart';
 import '../../version.dart';
-import 'command_status.dart';
+import '../models/command_status.dart';
 import 'create_base.dart';
 
 /// This class extends the flutter create abstraction,
@@ -151,14 +151,14 @@ Happy coding!
         CachePackage(templatePackageName, projectDir);
         if (boolArg('pub')!) {
           try {
-            await Flutter.pubGet(directory: projectDir);
+            await FlutterCli.pubGet(directory: projectDir);
           } catch (e) {
             _logger.w('Unable to pub get in ${projectDir.path}');
           }
         }
         return;
       } catch (e) {
-        await Flutter.pubAdd(
+        await FlutterCli.pubAdd(
           '$templatePackageName:$packageVersion',
           directory: projectDir,
         );
