+++ lib/src/config/android_config.dart
@@ -1,0 +1,7 @@
+/// The default android configuration for an at_app application
+const int targetSdkVersion = 30;
+const int minSdkVersion = 24;
+const int compileSdkVersion = 31;
+
+const String androidR8 = 'android.enableR8';
+const String androidR8Value = 'true';
