+++ lib/src/models/command_status.dart
@@ -1,0 +1,3 @@
+/// An enum for the status of the command
+/// The command can be a success, provide a warning, or completely fail.
+enum CommandStatus { success, warning, fail }
