+++ lib/src/models/exceptions/android_build_exception.dart
@@ -1,0 +1,4 @@
+/// An exception thrown when the android configuration fails to update
+class AndroidBuildException implements Exception {
+  final String message = 'Unable to update the android build configuration.';
+}
