+++ lib/src/models/exceptions/env_exception.dart
@@ -1,0 +1,4 @@
+/// An exception thrown when the environment variables fail to set
+class EnvException implements Exception {
+  final String message = 'Unable to update the environment variables.';
+}
