+++ lib/src/models/exceptions/package_exception.dart
@@ -1,0 +1,7 @@
+/// An exception thrown when a package is not found in the pub cache
+class PackageException implements Exception {
+  final String message;
+
+  PackageException(String packageName)
+      : message = 'No version of $packageName found in the pub cache.';
+}
