+++ lib/src/models/exceptions/template_exception.dart
@@ -1,0 +1,4 @@
+/// An exception thrown when a template file fails to be copied
+class TemplateException implements Exception {
+  final String message = 'Unable to replace project with the template.';
+}
