--- lib/src/util/cli/cli_base.dart
@@ -1,40 +1,0 @@
-import 'dart:io' show Process, ProcessException, ProcessResult;
-
-/// An abstraction layer to run commands using dart:io Process.
-class Cli {
-  /// Runs the specified [cmd] with the provided [args].
-  static Future<ProcessResult> run(
-    String cmd,
-    List<String> args, {
-    bool throwOnError = true,
-    String? directory,
-  }) async {
-    final result = await Process.run(cmd, args,
-        workingDirectory: directory, runInShell: true);
-
-    if (throwOnError) {
-      _throwIfProcessFailed(result, cmd, args);
-    }
-    return result;
-  }
-
-  static void _throwIfProcessFailed(
-    ProcessResult pr,
-    String process,
-    List<String> args,
-  ) {
-    if (pr.exitCode != 0) {
-      final values = {
-        'Standard out': pr.stdout.toString().trim(),
-        'Standard error': pr.stderr.toString().trim()
-      }..removeWhere((k, v) => v.isEmpty);
-
-      var message = 'Unknown error';
-      if (values.isNotEmpty) {
-        message = values.entries.map((e) => '${e.key}\n${e.value}').join('\n');
-      }
-
-      throw ProcessException(process, args, message, pr.exitCode);
-    }
-  }
-}
