--- lib/src/util/exceptions/android_build_exception.dart
@@ -1,4 +1,0 @@
-/// An exception thrown when the android configuration fails to update
-class AndroidBuildException implements Exception {
-  final String message = 'Unable to update the android build configuration.';
-}
