--- lib/src/util/exceptions/flutter_exception.dart
@@ -1,6 +1,0 @@
-/// An exception thrown when the Flutter cli exits with a non-zero code
-class FlutterException implements Exception {
-  final String message;
-
-  FlutterException(this.message);
-}
