+++ lib/src/util/file/env_manager.dart
@@ -1,6 +1,6 @@
 import 'dart:io' show Directory;
 
-import 'file_manager.dart';
+import '../../models/file_manager.dart';
 
 class EnvManager extends FileManager {
   EnvManager(Directory projectDir) : super(projectDir, '.', '.env');
