+++ lib/src/util/file/gitignore_manager.dart
@@ -1,6 +1,6 @@
 import 'dart:io' show Directory;
 
-import 'file_manager.dart';
+import '../../models/file_manager.dart';
 
 class GitignoreManager extends FileManager {
   GitignoreManager(Directory projectDir) : super(projectDir, '.', '.gitignore');
